[svn] Made the selection of a root deterministic.
The selection of a root during the uppass was not a deterministic process. The
edges where collected in a randomized pass, and then the first with minimum cost
was selected as the tree root. This has been modified by performing an initial
sort of the edges and then selecting the first one with minimum cost.