[svn] Improved the fuse () algorithm for better results and performance.
The fuse algorithm picked a pair of subtrees with the same terminals uniformly
at random and exchanged them without checking if they really correspond to
exactly the same tree. We now verify that they to correspond to different
trees, and improve the overall performance of the algorithm by removing some
useless diagnosis operations.