[svn] Modified the representation of sequences on the C side.
The C representation of sequences involved a complex scheme of pools of
sequences to improve the application performance. This was largely due to the
memory fragmentation caused by the repeated allocation of sequences.

The OCaml compiler has now a better handling of the memory allocation process,
so we have rolled a simpler representation that leaves the full responsability
of memory management on the OCaml Garbage Collector.