[svn] Fixed Issue 51 and unified code assignment for nodes and characters.
The assignment of codes to the characters and nodes was incremental using a
univeral counter. This added the requirement of transforming trees between
different processes when running in parallel. This has now been eliminated,
making the parallelization more efficient than before. This change also
simplifies the fix of issue 51.

Bugfix:
    Symptom:
        Issue 51 (loucrow and Fernando Marquez).

    Problem:
        The GatherTrees command had some bugs in the way it was merging trees
        from different processes.

    Solution:
        Simply merge the stored_trees and trees fields from
        Scripting.run and do not postprocess trees thanks to the change
        described above.