[svn] Fixed bug in the tree convertion from a forest input.
When reading an input tree that contained a forest, the code assignment of the
interior vertices was incorrect. This bug was introduced after the unification
of the interior codes (instead of using a universal counter of consecutive
codes).

The conversion function assumed that it was possible to assign all the codes in
each component independently, which isn't true once there is no universal
counter to follow. Instead, all the components need to be measured in size to
establish the codes that should be used on each.