[svn] Added the search () command. This is still experimental and remains undocumented
Documentation will be added when the experimental step finishes. For now I need
to be able to keep track of the versions during experimentation. The command
only works sequentially, and takes the optional arguments:

time:FLOAT (seconds for the search to be performed)
hits:INT (number of times the best tree should be found before finishing the
search)
memory:INT (the memory in number of words that can be used at most. This is
estimated, and not fully enforced.)