[svn] Added `Partition `Sets to programatically partition the data; modified search.
`Partiton `Sets x is a poyExtension only command to allow for partitioning
the trees during search in a programmer-provided set. This is very useful to
simplify the code for complex scripts.

The search command now takes advantage of this during the first round of
exhaustive_do searches by picking the first three shortest trees and the current
tree to produce the set of constraints.