[svn] Fixed missing codes in to_list function and made it pure OCaml.
During parallel execution random crashes could occur. These where caused by the
to_list function which was written in C but could break some invariants in the
Garbage Collector. This was fixing by making the an OCaml one.

A bug introduced during the change of representation of non-additive characters
was corrected. In this bug, the code of the characters where missing, producing
empty sets of characters in the *filter* functions. The codes are now properly
included.