[svn] Improved the behavior of X in aa sequences and cleaned c code.
The c code contained some unused variables that have been cleaned up.

The aminoacid cost matrices have now a special treatment for X,
and assign it cost 0 when aligned to everyone else and in addition its median to
any other element in the alphabet is itself. This option required a new
parameter in the cost matrices (the all elements code), and the proper selection
when filling the cost matrix and the median matrix.