[svn] Improved Search Status message.
The search status message said how many hits of a given tree where found. This
message was confusing for many users after the search () command which counts
total number of hits. To avoid it, it has been modified to account for the
number of trees in memory with the smallest cost.