Added Lexing rules for all the TNT commands as of July 2008.
Mesquite outputs the quote command, which we don't interpret. However, we
interpret all prefixes of size at least two for the commands that we can handle.
However, there is the danger of accepting a prefix that could have two meanings
by mistake. For this reason, there are two improvements in this commit:

- When adding a prefix to the Hennig lexer we now verify that the prefix if
  unique. If not, not only we don't add it, but we remove the clashing prefix.
- Added all the comands as listed in TNT's help to the lexing rules. We can't
  parse them, nevertheless, we will include them in the lexing rules. This also
  simplifies the process of adding new commands to interpret.

Given that the size of the lexing table is now large, I have replaced the
association table, with a hash table to go from quadratic to linear time.