Added rules to properly compile using OCaml 3.11.0+beta1
Added rules in the configure file to recognize 3.11.* as a valid OCaml version
number.

Added rules to select whether or not dynlink.cmxa is available (only in 3.11),
as it is needed by camlp4fulllib.cmxa in version 3.11.

Added the USE_NATIVEDYNLINK macro to the camlp4 preprocessor so that we can now
easily add support for plugins in the native version of POY and selectively
compile them if the application is being compiled using OCaml 3.11 or later.