Multiple improvements in the Makefile and compilation rules.
Splitted the depend generation in individual files depending on each source
file, and unified the compilation rules to use mostly the camlp4 preprocessor in
the compilation process. This change implied changing some variable names in a
few modules to eliminate clashes with Camlp4.

This new set of compilation rules should eliminate the need to make depend
completely, and eliminate incomplete dependencies.