If the synonyms file is not found, POY stops the script execution.
Julian Faivovich reported an undesirable behavior in POY, by which the failure
to find a synonym file does not cause the immediate termination of the program
execution. Though this was the initial intended behavior, this is clearly not
the best, and so, this change ensures that the exception is passed to the top
level of the program.