The characters created by static_approx keep the original sequence names.
When reading an input file, say chel.aln, POY names the characters included in
it as chel.aln:0, chel.aln:1, ... However, after transform (static_approx), the
name of the characters corresponding to the implied alignment become
ImpliedAlignment:0 .... This was messy.

We now maintain the name of the original character, and the resulting elements
of the implied alignment are chel.aln:0:ia:0, chel.aln:0:ia:1, .... making
easier for the users to trace back the origin of each change.