Ensure that the order in which synonyms are loaded is the same as the input file
The last change improved the error handling in synonym files. Unfortunately,
this meant that when loading a file that had some inconsistencies, the selected
synonym would depend in the order in which the contents of the file where
loaded.

To ensure it, I replaced a Hash table with a list, that has exactly the same
order of the input file. In that way, the processing will be predictable for the
end users. This change required the addition of
Parser.Dictionary.of_channel_assc.