Addded the standard_tabu_searcher queue manager.
The standard_tabu_searcher takes a set of negative constraints (a la bremer
support), and these are maintained as sets of clades that should not occur in a
tree during the search.

The new tabu searcher queue manager applies a subset of the tabus passed in
rounds. This new searcher has not been tested and has not been plugged to the
rest of the code, but this commit adds the necessary class to lay the foundation
to properly plug it in.