Added the print_newick plugin, to print newick trees in graphical format.
The added value of this plugin is that no data need to be loaded in POY to be
able to perform this operation. Notice that this plugin accepts branch lengths
in newick format. The POY 4 parsers do not support this kind of format (only
newick trees with no branch lengths or using the POY branch length format).