Eliminated the reference used in Poy at runtime to maintain the program state.
Poy.script maintained the state of the program. Unfortunately,
Phylo.get_console_run and Phylo.set_console_run provided means to change the
state stored in the Phylo module. This caused an inconsistency, as the use of
scripting within OCAML could change the state of the program, but not the state
that could be extracted with the Phylo module.

I have now replaced Poy.script with corresponding occurrences of
Phylo.get_console_run and Phylo.set_console_run when necessary.